Renamed parameters (#25)

+++ src/main/java/com/example/demo/service/StudentService.java
@@ -5,12 +5,10 @@ import java.util.List;
 import com.example.demo.entity.Student;
 
 public interface StudentService {
-
     public void addNewStudent(Student student);
-    public void deleteStudent(Long studentId);
-    public void updateStudent(Long studentId, String name, String email);
-    public Student getStudentById(Long studentId);
+    public void deleteStudent(int studentId);
+    public void updateStudent(int studentId, String name, String email);
+    public Student getStudentById(int studentId);
     public Student getStudentByEmail(String email);
     public List<Student> getStudents();
-
 }
